requirements.txt
requests>=2.28.0
Pillow>=9.0.0
google-api-python-client>=2.88.0
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
python-dotenv>=1.0.0
pytz>=2023.3
